change: moving to deprecated  dropdown & social Share
