+++ CHANGELOG.mdx
@@ -20,7 +20,6 @@ Prefix the change with one of these keywords:
 
 - Upgrade footer and body classes to display footer at page bottom
 
-
 ## [0.9.8]
 
 ### Added
