+++ lib/data/DropDownData.ts
@@ -1,4 +1,4 @@
-import { DropDownItemProps } from '../components/DropDown/DropDown'
+import { DropDownItemProps } from '../_deprecated/DropDown/DropDown'
 
 export const DropDownItemData: DropDownItemProps[] = [
   {
