+++ lib/main.ts
@@ -10,7 +10,6 @@ export { ButtonGroup } from './components/ButtonGroup/ButtonGroup'
 export { Calendar } from './components/Calendar/Calendar'
 export { Card } from './components/Card/Card'
 export { Description } from './components/Description/Description'
-export { DropDown } from './components/DropDown/DropDown'
 export { Embed } from './components/Embed/Embed'
 export { Filter } from './components/Filter/Filter'
 export { FooterBasic } from './components/Footer/FooterBasic/FooterBasic'
@@ -22,7 +21,7 @@ export { Icon } from './components/Icon/Icon'
 export { LinkProvider } from './components/LinkProvider'
 export { Listing } from './components/Listing/Listing'
 export { Location } from './components/Location/Location'
-// export { LocationPicker } from './components/LocationPicker/LocationPicker'
+// TODO: bug fix export { LocationPicker } from './components/LocationPicker/LocationPicker'
 export { Meta } from './components/Meta/Meta'
 export { Modal } from './components/Modal/Modal'
 export { MultiDayCalendar } from './components/MultiDayCalendar/MultiDayCalendar'
@@ -32,7 +31,6 @@ export { Pagination } from './components/Pagination/Pagination'
 export { Quote } from './components/Quote/Quote'
 export { Search } from './components/Search/Search'
 export { SearchForm } from './components/SearchForm/SearchForm'
-export { SocialShare } from './components/SocialShare/SocialShare'
 export { Table } from './components/Table/Table'
 export { TextImage } from './components/TextImage/TextImage'
 export { Toast } from './components/Toast/Toast'
